+++ CHANGELOG.md
@@ -4,3 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.0] - 2024-12-10
+- Initial release
+
+## [2.0.0] - 2025-01-27
+
+### Added
+- Added some documentation and examples.
+
+### Changed
+- Changed the way tools are defined. Now tools are functors.
+
+### Removed
+- Removed old tools implementation.
+
+### Fixed
+- Fixed agent handoff capability.